refactor: Product 및 HistoryEntity id 타입 Long으로 변경

+++ app/src/main/java/woowacourse/shopping/data/remote/cart/CartRequest.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CartRequest(
-    val productId: Int,
-    val quantity: Int
+    val productId: Long,
+    val quantity: Int,
 )
